pyproject.toml
[tool.poetry]
name = "tnt-python"
version = "0.1.0"
description = "Telegram notification tool"
license = "MIT"
maintainers = ["axeman <axeman.ofic@gmail.com>"]
authors = ["axeman <axeman.ofic@gmail.com>"]
repository = "https://github.com/watchmanix/pytnt"
keywords = ["python", "telegram"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
]
readme = "README.md"
packages = [{include = "tnt_python"}]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.23.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.2"
python-dotenv = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
